introduce support for the SolidRun clearfog
Initial support for the SolidRun Clearfog laned in FreeBSD with r319914.
The board has acceptable support in mainline U-Boot, as well as a
vendor fork based on 2018.01.
For now an MBR partition table is used, with a fat partition to load
DTB and ubldr from, reserved space covering the first 1M of the disk
to make sure installing u-boot later doesn't destroy the boot partition.

Signed-off-by: Josua Mayer <josua.mayer97@gmail.com>